Coverity fix for y2k38_issue

+++ src/rtmessage/rtrouteBase.c
@@ -289,7 +289,7 @@ _rtdirect_dispatch_message_from_client(rtConnectedClient* clnt, rtRouteEntry* pD
     if((clnt->header.flags & rtMessageFlags_Request))
     {
       rtTime_Now(&ts);
-      clnt->header.T3 = ts.tv_sec;
+      clnt->header.T3 = (uint64_t)ts.tv_sec;
     }
 #endif
     rtLog_Debug("DispatchMessage topic=%s expression=%s", clnt->header.topic, route->expression);
@@ -339,7 +339,7 @@ static rtError
 rtConnectedClient_Read(rtConnectedClient* clnt, rtRouteEntry* myDirectRoute)
 {
   ssize_t bytes_read;
-  int bytes_to_read = (clnt->bytes_to_read - clnt->bytes_read);
+  size_t bytes_to_read = (clnt->bytes_to_read - clnt->bytes_read);
 
 #ifdef MSG_ROUNDTRIP_TIME
   rtTime_t daemon_request = {0};
@@ -347,12 +347,12 @@ rtConnectedClient_Read(rtConnectedClient* clnt, rtRouteEntry* myDirectRoute)
   if(clnt->header.flags & rtMessageFlags_Request)
   {
      rtTime_Now(&daemon_request);
-     clnt->header.T2 = daemon_request.tv_sec;
+     clnt->header.T2 = (uint64_t)daemon_request.tv_sec;
   }
   if(clnt->header.flags & rtMessageFlags_Response)
   {
      rtTime_Now(&daemon_response);
-     clnt->header.T5 = daemon_response.tv_sec;
+     clnt->header.T5 = (uint64_t)daemon_response.tv_sec;
   }
 #endif
 
@@ -412,7 +412,7 @@ rtConnectedClient_Read(rtConnectedClient* clnt, rtRouteEntry* myDirectRoute)
 #endif
         clnt->bytes_to_read += clnt->header.payload_length;
         clnt->state = rtConnectionState_ReadPayload;
-        int incoming_data_size = clnt->bytes_to_read + clnt->bytes_read;
+        uint64_t incoming_data_size = clnt->bytes_to_read + clnt->bytes_read;
         if(clnt->read_buffer_capacity < incoming_data_size)
         {
           uint8_t * ptr = (uint8_t *)rt_try_realloc(clnt->read_buffer, incoming_data_size);
@@ -420,11 +420,11 @@ rtConnectedClient_Read(rtConnectedClient* clnt, rtRouteEntry* myDirectRoute)
           {
             clnt->read_buffer = ptr;
             clnt->read_buffer_capacity = incoming_data_size;
-            rtLog_Info("Reallocated read buffer to %d bytes to accommodate traffic.", incoming_data_size);
+            rtLog_Info("Reallocated read buffer to %ld bytes to accommodate traffic.", incoming_data_size);
           }
           else
           {
-            rtLog_Info("Couldn't not reallocate read buffer to accommodate %d bytes. Message will be dropped.", incoming_data_size);
+            rtLog_Info("Couldn't not reallocate read buffer to accommodate %ld bytes. Message will be dropped.", incoming_data_size);
             _rtConnection_ReadAndDropBytes(clnt->fd, clnt->header.payload_length);
             rtConnectedClient_Reset(clnt);
             break;
